added authentication and return more user info for login - votes controller not done yet

+++ app/controllers/users_controller.rb
@@ -12,34 +12,47 @@ class UsersController < ApplicationController
   end
 
   def create
-    user = User.new(user_params)
+    if authenticate_token?(params[:auth_token])
+      user = User.new(user_params)
 
-    if user.save
-      session[:user_id] = user.id
-      render json: { user: user, status: 201 }
+      if user.save
+        session[:user_id] = user.id
+        render json: { user: user, status: 201 }
+      else
+        render json: user.errors, status: 422
+      end
     else
-      render json: user.errors, status: 422
+      render json: { message: "You must be logged in to do that", status: 401}
     end
   end
 
   def update
-    user = set_user
-
-    if user.update(user_params)
-      render json: user
+    if user = set_user
+      if authenticate_ownership?(params[:auth_token], user)
+        if user.update(user_params)
+          render json: user
+        else
+          render json: user.errors, status: 422
+        end
+      else
+        render json: { message: "You must be logged in to do that", status: 401 }
+      end
     else
-      render json: user.errors, status: 422
+      render json: { message: "User does not exist", status: 401 }
     end
   end
 
   def destroy
-    user = set_user
-    # if user.id == current_user.id
-      user.destroy
-      render json: { message: "User was successfully deleted." }
-    # else
-      # render json: { status: 404 }
-    # end
+    if user = set_user
+      if authenticate_ownership?(params[:auth_token], user)
+        user.destroy
+        render json: { message: "User was successfully deleted." }
+      else
+        render json: { message: "You must be logged in to do that", status: 401 }
+      end
+    else
+      render json: { message: "User does not exist", status: 401 }
+    end
   end
 
   private
